完成基础框架梳理，包括菜单管理bug，商户管理，用户管理等

+++ trace-system-common/src/main/java/com/ejunhai/trace/merchant/model/MerchantInfo.java
@@ -10,7 +10,7 @@ public class MerchantInfo implements Serializable {
 
     private String businessLine;
 
-    private Integer merchantLevel;
+    private Byte merchantLevel;
 
     private Date openTime;
 
@@ -62,11 +62,11 @@ public class MerchantInfo implements Serializable {
         this.businessLine = businessLine == null ? null : businessLine.trim();
     }
 
-    public Integer getMerchantLevel() {
+    public Byte getMerchantLevel() {
         return merchantLevel;
     }
 
-    public void setMerchantLevel(Integer merchantLevel) {
+    public void setMerchantLevel(Byte merchantLevel) {
         this.merchantLevel = merchantLevel;
     }
